refactor: Change Coord coordinates from i32 to i8 for improved memory efficiency

+++ src/game_objects.rs
@@ -328,8 +328,8 @@ pub struct Battlesnake {
 /// * `y` - The y-coordinate of the Coord. Example: 5
 #[derive(Deserialize, Serialize, Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct Coord {
-    pub(super) x: i32,
-    pub(super) y: i32,
+    pub(super) x: i8,
+    pub(super) y: i8,
 }
 
 impl Coord {
